# openAPI_documents/market-openAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "market-api",
    "description": "Market application API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/THeikki/market-api/1.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/user/register": {
      "post": {
        "description": "Register new user",
        "operationId": "post-register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "201": {
            "description": "New user registered"
          },
          "400": {
            "description": "validate: errors"
          },
          "409": {
            "description": "User already registered"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "description": "Login user",
        "operationId": "post-login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successfull"
          },
          "400": {
            "description": "validate: errors"
          },
          "401": {
            "description": "Invalid password"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "title": "User",
        "required": [
          "Pwd",
          "UserName"
        ],
        "type": "object",
        "properties": {
          "UserName": {
            "type": "string"
          },
          "Pwd": {
            "type": "string"
          }
        },
        "x-examples": {
          "example user": {
            "UserName": "Seller1",
            "Pwd": "password"
          }
        }
      },
...
    }
  }
}